refactor(codegen): inline reflectapi-schema-codegen as an internal module + 0.17.2-alpha.2
The reflectapi-schema-codegen crate had exactly one consumer
(reflectapi/src/codegen/python.rs) and wasn't re-exported through
reflectapi's public API. Publishing it would commit us to a semver
contract on the still-evolving semantic IR (#129), and the first
release hit the crates.io Trusted Publishing limitation that new
crates can't be created via OIDC.

Move ids/normalize/semantic/symbol back into the reflectapi crate as
reflectapi::codegen::schema (pub(crate) only). The architectural
separation that PR #128 established is preserved at the module
boundary; SymbolId / SchemaIds / SemanticSchema can still grow without
touching reflectapi-schema (the wire/interchange crate).

Concretely:

- New: reflectapi/src/codegen/schema/{mod,ids,normalize,semantic,symbol}.rs
- Drop: reflectapi-schema-codegen crate + workspace member
- Drop: reflectapi's optional dep on it + the codegen feature reference
- python.rs: 'use reflectapi_schema_codegen as schema_codegen'
            -> 'use crate::codegen::schema as schema_codegen'
- Module-level #![allow(dead_code)] inside schema/ — most accessor
  methods are exercised only by the module's own tests today and stay
  as the supported IR shape for future backend reuse.
- mod.rs re-exports trimmed to what python.rs actually consumes; the
  rest is reachable via self::ids::*, self::normalize::*, etc.
- release.yml drops the schema-codegen publish step and version-check
  entry; README's release section rewritten to match the workflow
  (Trusted Publishing, dual SemVer/PEP 440 form, tag-driven).

Versions: 0.17.2-alpha.1 -> 0.17.2-alpha.2 (Cargo) / 0.17.2a2 (PEP 440).
