+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "reflectapi-demo/clients/rust/generated",
     "reflectapi-derive",
     "reflectapi-schema",
-    "reflectapi-schema-codegen",
 ]
 resolver = "2"
