+++ reflectapi-derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reflectapi-derive"
-version = "0.17.2-alpha.1"
+version = "0.17.2-alpha.2"
 edition = "2021"
 
 license = "Apache-2.0"
@@ -22,7 +22,7 @@ workspace = true
 proc-macro = true
 
 [dependencies]
-reflectapi-schema = { path = '../reflectapi-schema', version = "0.17.2-alpha.1" }
+reflectapi-schema = { path = '../reflectapi-schema', version = "0.17.2-alpha.2" }
 
 proc-macro2 = "1.0"
 quote = "1.0"
