+++ reflectapi-schema/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reflectapi-schema"
-version = "0.17.2-alpha.1"
+version = "0.17.2-alpha.2"
 edition = "2021"
 
 license = "Apache-2.0"
