+++ reflectapi/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reflectapi"
-version = "0.17.2-alpha.1"
+version = "0.17.2-alpha.2"
 edition = "2021"
 
 license = "Apache-2.0"
@@ -20,9 +20,8 @@ workspace = true
 
 [dependencies]
 # workspace dependencies
-reflectapi-derive = { path = "../reflectapi-derive", version = "0.17.2-alpha.1" }
-reflectapi-schema = { path = "../reflectapi-schema", version = "0.17.2-alpha.1" }
-reflectapi-schema-codegen = { path = "../reflectapi-schema-codegen", version = "0.17.2-alpha.1", optional = true }
+reflectapi-derive = { path = "../reflectapi-derive", version = "0.17.2-alpha.2" }
+reflectapi-schema = { path = "../reflectapi-schema", version = "0.17.2-alpha.2" }
 
 # mandatory 3rd party dependencies
 serde = { version = "1.0.197", features = ["derive"] }
@@ -84,7 +83,6 @@ codegen = [
     "dep:indexmap",
     "dep:check_keyword",
     "dep:serde_json",
-    "dep:reflectapi-schema-codegen",
 ]
 rt = ["dep:http", "dep:serde_json", "dep:bytes", "dep:url"]
 glob = ["reflectapi-schema/glob"]
